Rename task alert store to alert store and move task alert component into tasks folder
